+++ src/App.vue
@@ -55,7 +55,7 @@ const pageListener = (
             sendResponse({
               error: true,
               message:
-                "ClearWallet: Invalid PAGE request method " + (message?.method ?? ""),
+                "MetatarzWallet: Invalid PAGE request method " + (message?.method ?? ""),
             });
             break;
           }