+++ src/views/HomeTab.vue
@@ -4,27 +4,24 @@
       <ion-toolbar>
         <ion-title>
           <ion-avatar
-            @click="openTab('https://clear-wallet.flashsoft.eu/docs/')"
+            @click="openTab('https://wallet.metatarz.xyz')"
             class="link-docs"
             style="margin: 0.3rem; width: 1.6rem; height: 1.6rem; display: inline-flex"
           >
-            <img alt="clw" :src="getUrl('assets/extension-icon/wallet_32.png')" />
+            <img alt="clw" :src="getUrl('assets/extension-icon/wallet_48.png')" />
           </ion-avatar>
           <span
-            @click="openTab('https://clear-wallet.flashsoft.eu/docs/')"
+            @click="openTab('https://wallet.metatarz.xyz')"
             class="link-docs"
             style="position: absolute; top: 0.35rem; margin-left: 0.3rem"
           >
             <span style="font-size: 0.9rem; font-weight: bold; color: #aca3bb">
-              Clear
+              Metatarz
             </span>
             <span style="font-size: 0.9rem; color: #aca3bb"> Wallet </span>
           </span>
           <span
             v-if="version"
-            @click="
-              openTab('https://clear-wallet.flashsoft.eu/docs/automated-changelog/')
-            "
             style="
               position: absolute;
               right: 1.1rem;
@@ -39,7 +36,7 @@
           >
           <span
             class="github-icon"
-            @click="openTab('https://github.com/andrei0x309/clear-wallet/')"
+            @click="openTab('https://github.com/metatarzxyz/metatarz-wallet/')"
             ><GitHub
           /></span>
         </ion-title>